requirements.txt
notion-client
openai
google-api-python-client
google-auth-oauthlib
python-dotenv
discord.py
sounddevice
numpy
scipy
websockets
pyaudio
pyyaml
